+++ src/app/(house)/search/page.tsx
@@ -1,4 +1,4 @@
-import { Filter } from "@/features/main/components/Filter";
+import { Filters } from "@/features/main/components/Filters";
 import { SearchResultList } from "@/features/search/components/SearchResultList";
 import { Suspense } from "react";
 
@@ -6,7 +6,7 @@ export default function Search() {
   return (
     <main>
       <Suspense fallback={<></>}>
-        <Filter />
+        <Filters />
         <SearchResultList />
       </Suspense>
     </main>
